# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: send-flow API Blueprint
  version: 0.1.0
  description: >
    Contract-first OpenAPI specification derived from docs/api/* only.
    This file documents the product blueprint for send-flow and may include
    endpoints that are not yet implemented in code.
servers:
  - url: /
...
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Bearer token used by dashboard sessions and workspace API keys depending
        on the endpoint. Endpoints that already contain `workspace_id` in the
        path remain stateless by path, and dashboard-session clients should
        provide workspace context explicitly per request when needed.
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: WorkspaceApiKey
      description: Workspace API key sent as a bearer credential, for example `Bearer sk_live_xxx`.
    providerSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: Provider-specific webhook signature header.
...
paths:
...
  /api/events/stream:
    get:
      tags: [Operations]
      operationId: streamEvents
      x-implementation-status: implemented
      summary: Stream runtime events over SSE
      responses:
        '200':
          description: SSE stream opened
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: connected
                  data: send-flow stream connected

                  : keepalive
        '500':
          description: Streaming transport is not supported
...